backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
pymysql
pymongo
django-cors-headers
dotenv
cryptography